# backend/gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API документация для budget-app",
        "title": "Budget API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "JSON",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.LoginHandlerIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.LoginHandlerOut"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJSON"
                        }
                    }
                }
            }
        },
...
        "/auth/refresh": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh tokens",
                "parameters": [
                    {
                        "description": "JSON",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.RefreshHandlerIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.TokensOutDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJSON"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register",
                "parameters": [
                    {
                        "description": "JSON",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.RegisterHandlerIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJSON"
                        }
                    }
                }
            }
        },
...
        "/ledger/budgets": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ledger"
                ],
                "summary": "List budgets",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Фильтр по дате ОТ в формате 2025-01 (год-месяц)",
                        "name": "period_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по дате ДО в формате 2025-01 (год-месяц)",
                        "name": "period_to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ledger.BudgetListHandlerOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJSON"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ledger"
                ],
                "summary": "Add budget",
                "parameters": [
                    {
                        "description": "JSON",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ledger.BudgetAddHandlerInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ledger.BudgetAddHandlerOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJSON"
                        }
                    }
                }
            }
        },
        "/ledger/budgets/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ledger"
                ],
                "summary": "Get budget",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ledger.BudgetGetHandlerOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJSON"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "ledger"
                ],
                "summary": "Delete budget",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJSON"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ledger"
                ],
                "summary": "Patch budget",
                "parameters": [
                    {
                        "description": "JSON",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ledger.BudgetPatchHandlerInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ledger.BudgetPatchHandlerOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJSON"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "auth.AccountOutDTO": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "profileName": {
                    "type": "string"
                },
                "profileSurname": {
                    "type": "string"
                }
            }
        },
        "auth.LoginHandlerIn": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 0
                }
            }
        },
        "auth.LoginHandlerOut": {
            "type": "object",
            "properties": {
                "accout": {
                    "$ref": "#/definitions/auth.AccountOutDTO"
                },
                "tokens": {
                    "$ref": "#/definitions/auth.TokensOutDTO"
                }
            }
        },
        "auth.RefreshHandlerIn": {
            "type": "object",
            "properties": {
                "refreshJWT": {
                    "type": "string"
                }
            }
        },
        "auth.RegisterHandlerIn": {
            "type": "object",
            "required": [
                "email",
                "profileName",
                "profileSurname"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 0
                },
                "profileName": {
                    "type": "string"
                },
                "profileSurname": {
                    "type": "string"
                }
            }
        },
        "auth.TokensOutDTO": {
            "type": "object",
            "properties": {
                "accessJWT": {
                    "type": "string"
                },
                "refreshJWT": {
                    "type": "string"
                }
            }
        },
        "ledger.BudgetAddHandlerInput": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "categoryID": {
                    "type": "integer"
                },
                "period": {
                    "$ref": "#/definitions/ledger.BudgetAddHandlerInputPeriod"
                }
            }
        },
        "ledger.BudgetAddHandlerInputPeriod": {
            "type": "object",
            "properties": {
                "month": {
                    "type": "integer"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "ledger.BudgetAddHandlerOutput": {
            "type": "object",
            "properties": {
                "item": {
                    "$ref": "#/definitions/ledger.BudgetOutput"
                }
            }
        },
        "ledger.BudgetGetHandlerOutput": {
            "type": "object",
            "properties": {
                "hitCache": {
                    "type": "boolean"
                },
                "item": {
                    "$ref": "#/definitions/ledger.BudgetOutput"
                }
            }
        },
        "ledger.BudgetListHandlerOutput": {
            "type": "object",
            "properties": {
                "hit_cache": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ledger.BudgetOutput"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "ledger.BudgetOutput": {
            "type": "object",
            "properties": {
                "accountID": {
                    "type": "string"
                },
                "amount": {
                    "type": "string"
                },
                "categoryID": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "period": {
                    "$ref": "#/definitions/ledger.BudgetOutputPeriod"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "ledger.BudgetOutputPeriod": {
            "type": "object",
            "properties": {
                "month": {
                    "type": "integer"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "ledger.BudgetPatchHandlerInput": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "categoryID": {
                    "type": "integer"
                },
                "period": {
                    "$ref": "#/definitions/ledger.BudgetAddHandlerInputPeriod"
                }
            }
        },
        "ledger.BudgetPatchHandlerOutput": {
            "type": "object",
            "properties": {
                "item": {
                    "$ref": "#/definitions/ledger.BudgetOutput"
                }
            }
        },
...
        "middleware.ErrorJSON": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "hints": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "textCode": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}