# backend/gateway/internal/delivery/http/controller/v1/auth/dto.go
package auth

type AccountOutDTO struct {
	ID             string `json:"id"`
	Email          string `json:"email"`
	ProfileName    string `json:"profileName"`
	ProfileSurname string `json:"profileSurname"`
}

type TokensOutDTO struct {
	RefreshJWT string `json:"refreshJWT"`
	AccessJWT  string `json:"accessJWT"`
}